refactor: move database error
