--- packages/apps/reputation-oracle/server/src/common/errors/manifest.ts
@@ -1,9 +1,0 @@
-import { BaseError } from './base';
-
-class ManifestError extends BaseError {}
-
-export class MissingManifestUrlError extends ManifestError {
-  constructor(readonly escrowAddress: string) {
-    super('Manifest url is missing for escrow');
-  }
-}
