+++ packages/apps/reputation-oracle/server/src/database/index.ts
@@ -1,3 +1,8 @@
 export { BaseEntity } from './base.entity';
 export { BaseRepository } from './base.repository';
 export { DatabaseModule } from './database.module';
+export {
+  DatabaseErrorMessages,
+  DatabaseError,
+  isDuplicatedError,
+} from './errors';
